why did I use int* (fix font effects corrupting when heavy memory load)

+++ common/Storyboarding/CommandValues/CommandDecimal.cs
@@ -16,7 +16,7 @@ namespace StorybrewCommon.Storyboarding.CommandValues
                 this.value = 0;
                 return;
             }
-            this.value = (decimal)value;
+            this.value = value;
         }
 
         public bool Equals(CommandDecimal other) => value.Equals(other.value);
